change usage of uniq in README
It doesn't make sense to call `uniq` on the array. If there are a page size of 10 with doubled results  and I remove them from the array the page has less then 10 items. If I call `uniq` on the relation the db handles this and I get the full page size of 10 items. And I can work with the relation in the frontend and I'm not limited to the array.